Renamed some methods and top level object.
